+++ packages/@raid-ui/control/range/src/range.stories.js
@@ -96,6 +96,9 @@ export const Basic = () => (
       width={0.75}
       height={8}
     />
+    <Range
+      disabled
+    />
   </Stack>
 )