+++ changelog.md
@@ -6,6 +6,7 @@ across minors, although they will be highlighted here with a :boom:.
 
 * _add_ sx to icon wrapper element
 * _update_ :boom: image props for passed components now accept the component (not a function)
+* _add_ range component disabled state
 
 ## Update guide